backend/requirements.txt
ansible==2.9.10
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
djangorestframework==3.11.0
django-url-filter==0.3.15
django-cors-headers==3.4.0
idna==2.9
psycopg2==2.8.5
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
django-rest-auth==0.9.5
django-allauth==0.42.0
djangorestframework-jwt==1.11.0
cython==0.29.20
